+++ src/main/kotlin/racingcar/model/InputValidator.kt
@@ -1,6 +1,6 @@
 package racingcar.model
 
-object CarNameValidator {
+object InputValidator {
 
     fun validateCarName(carNames: List<String>) {
         carNames.forEach { name ->
@@ -16,4 +16,9 @@ object CarNameValidator {
         return indexMap.filterValues { it.size > 1 }
     }
 
+    fun validateRaceCount(raceCount: String) {
+        val count = raceCount.toIntOrNull()
+        require(count != null) { "경주를 시도할 횟수는 숫자만 입력해주세요." }
+        require(count > 0) { "경주를 시도할 횟수는 1 이상이어야 합니다." }
+    }
 }
